total expense bug fix

+++ app/user/analyticsPage.tsx
@@ -106,9 +106,9 @@ const AnalyticsPage: React.FC<AnalyticsPageProp> = ({ userID }) => {
       if (!groupedData[key][category]) {
         groupedData[key][category] = 0;
       }
-      console.log(parseInt(expense.amount))
+      console.log(parseFloat(expense.amount))
   
-      groupedData[key][category] += parseInt(expense.amount);
+      groupedData[key][category] += parseFloat(expense.amount);
     });
   
     const labels: string[] = Object.keys(groupedData);
